Play with forks

+++ philo.h
@@ -6,7 +6,7 @@
 /*   By: ioleinik <ioleinik@student.42wolfsburg.de> +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/01 15:59:00 by ioleinik          #+#    #+#             */
-/*   Updated: 2021/09/25 16:03:59 by ioleinik         ###   ########.fr       */
+/*   Updated: 2021/09/25 16:48:44 by ioleinik         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,7 +46,7 @@ typedef struct s_arg
 	int				time_sleep;
 	int				num_eat;
 	int				finish;
-	unsigned long	start;
+	long long		start;
 	t_phil			*phils;
 	pthread_mutex_t	*forks;
 	pthread_mutex_t	access;
@@ -57,7 +57,7 @@ int				check_contract(t_arg *args, int argc, char **argv);
 int				ft_atoi(const char *str);
 int				ft_terror(char *s);
 void			m_print(t_arg *args, char *str, int id);
-unsigned long	get_time(void);
+long long		get_time(void);
 void			*routine(void *philosoph);
 
 #endif
